fix cycle counter tracker overflow

+++ apps/ipcbench/include/ipcbench.h
@@ -3,7 +3,7 @@
  * types start with an uppercase letter. So we typedef an unsigned
  * integer appropriately.
  */
-typedef unsigned int UnsignedInteger;
+typedef unsigned long long UnsignedInteger;
 
 #define NTRIALS             1000
 #define TRIAL_SPOT_INTERVAL 333
